# doc/api.yaml
openapi: 3.0.3
info:
  title: HFUT Graduation Project API
  description: 合肥工业大学毕业设计项目 API 接口文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://api.example.com/api/v1
    description: 生产环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 认证，格式：Bearer {token}
  schemas:
    Response:
      type: object
      properties:
        code:
          type: integer
          description: 状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: 操作成功
        data:
          type: object
          description: 响应数据
...
    LikeCreateRequest:
      type: object
      required:
        - ext_type
        - ext_id
      properties:
        ext_type:
          type: integer
          description: 关联类型，1:articles（文章） 2:comments（评论） 3:goods（商品）
          example: 1
        ext_id:
          type: integer
          description: 关联ID（文章ID、评论ID或商品ID）
          example: 1
    LikeResponse:
      type: object
      properties:
        is_liked:
          type: boolean
          description: 是否已点赞
          example: true
...
paths:
...
  /likes/toggle:
    post:
      tags:
        - 点赞
      summary: 切换点赞状态
      description: 切换点赞状态（点赞/取消点赞），需要认证
      operationId: toggleLike
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeCreateRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LikeResponse'
...